functions/requirements.txt
azure-functions
azure-storage-blob
pandas
openpyxl
requests
python-dotenv
